+++ db/models/university_category.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Table, ForeignKey, TIMESTAMP, text
+from sqlalchemy import Column, BIGINT, String, Table, ForeignKey, TIMESTAMP, text
 from sqlalchemy.orm import relationship
 from db.database import Base
 
@@ -6,15 +6,15 @@ from db.database import Base
 scholarship_university_category_table = Table(
     'scholarship_university_category',  # DB에 생성될 테이블 이름
     Base.metadata,
-    Column('scholarship_id', Integer, ForeignKey('scholarship.id'), primary_key=True),
-    Column('university_category_id', Integer, ForeignKey('university_category.id'), primary_key=True)
+    Column('scholarship_id', BIGINT, ForeignKey('scholarship.id'), primary_key=True),
+    Column('university_category_id', BIGINT, ForeignKey('university_category.id'), primary_key=True)
 )
 
 class UniversityCategory(Base):
     __tablename__ = "university_category"
     __table_args__ = {'extend_existing': True}
     
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BIGINT, primary_key=True, autoincrement=True)
 
     name = Column(String(255), nullable=False, unique=True, index=True)
